Extract Adjust and Audio tabs into Inspector/Tabs/
Move the Adjust (color/curves/wheels) and Audio tab logic out of the 1503-line InspectorView into Tabs/AdjustTab.swift and Tabs/AudioTab.swift as extensions; relocate TextTab/AIEditTab into Tabs/ for symmetry. Behavior-identical: same type/state, only source relocation + relaxing a few shared helpers from private to internal for cross-file access. Shell drops to ~1020 lines.

Video tab + the shared inspector kit (animatableRow/keyframeControls/propertyRow/collapsibleHeader/commitToClips) remain in the shell — they're interleaved, so splitting them cleanly is a separate refactor.
