+++ Sources/PalmierPro/Inspector/InspectorView.swift
@@ -24,7 +24,7 @@ struct InspectorView: View {
     }
 
     @State private var preferredTab: ClipTab = .video
-    @State private var adjustSubTab: AdjustTab = .basic
+    @State var adjustSubTab: AdjustTab = .basic
     @State private var preferredAssetTab: AssetTab = .details
     @State private var transformExpanded = true
 
@@ -192,7 +192,7 @@ struct InspectorView: View {
         return editor.mediaAssets.first { $0.id == clip.mediaRef }
     }
 
-    private var nonTextVisualClips: [Clip] {
+    var nonTextVisualClips: [Clip] {
         selectedVisualClips.filter { $0.mediaType != .text }
     }
 
@@ -282,7 +282,7 @@ struct InspectorView: View {
     }
 
     /// Capsule segmented control — used for the Adjust sub-tabs.
-    private func capsuleTabBar(titles: [String], selected: String?, onSelect: @escaping (String) -> Void) -> some View {
+    func capsuleTabBar(titles: [String], selected: String?, onSelect: @escaping (String) -> Void) -> some View {
         HStack(spacing: 0) {
             ForEach(titles, id: \.self) { title in
                 let isActive = selected == title
@@ -332,7 +332,7 @@ struct InspectorView: View {
         keyframesToggleBar(enabled: single != nil)
     }
 
-    private func keyframesToggleBar(enabled: Bool) -> some View {
+    func keyframesToggleBar(enabled: Bool) -> some View {
         let on = editor.keyframesPanelVisible
         return HStack {
             Spacer()
@@ -358,488 +358,7 @@ struct InspectorView: View {
     }
 
     @ViewBuilder
-    private func audioTabContent() -> some View {
-        let audios = selectedAudioClips
-        let single = audios.count == 1 ? audios.first : nil
-        let kfVisible = single != nil && editor.keyframesPanelVisible
-
-        if let clip = single, kfVisible {
-            HStack(alignment: .top, spacing: 0) {
-                VStack(alignment: .leading, spacing: AppTheme.Spacing.smMd) {
-                    // Match the kf panel's ruler+strip header height so Volume aligns with its lane.
-                    sectionTitleLabel(title: "Levels")
-                        .frame(height: KeyframesMetrics.headerHeight, alignment: .bottomLeading)
-                    volumeRow(audios: audios)
-                    fadeRow(label: "Fade In", clips: audios, edge: .left)
-                        .padding(.trailing, KeyframesMetrics.controlsColumnWidth + AppTheme.Spacing.sm)
-                    fadeRow(label: "Fade Out", clips: audios, edge: .right)
-                        .padding(.trailing, KeyframesMetrics.controlsColumnWidth + AppTheme.Spacing.sm)
-                    if nonTextVisualClips.isEmpty {
-                        speedSection(clips: audios)
-                            .padding(.trailing, KeyframesMetrics.controlsColumnWidth + AppTheme.Spacing.sm)
-                            .padding(.top, AppTheme.Spacing.md)
-                    }
-                }
-                .frame(maxWidth: .infinity, alignment: .leading)
-                .padding(.trailing, AppTheme.Spacing.sm)
-                Divider()
-                KeyframesPanel(clip: clip)
-                    .frame(maxWidth: .infinity, alignment: .leading)
-                    .padding(.leading, AppTheme.Spacing.sm)
-            }
-        } else {
-            VStack(alignment: .leading, spacing: AppTheme.Spacing.md) {
-                VStack(alignment: .leading, spacing: AppTheme.Spacing.smMd) {
-                    sectionTitleLabel(title: "Levels")
-                    volumeRow(audios: audios)
-                    fadeRow(label: "Fade In", clips: audios, edge: .left)
-                    fadeRow(label: "Fade Out", clips: audios, edge: .right)
-                }
-                if nonTextVisualClips.isEmpty {
-                    speedSection(clips: audios)
-                }
-            }
-        }
-
-        keyframesToggleBar(enabled: single != nil)
-    }
-
-    @ViewBuilder
-    private func volumeRow(audios: [Clip]) -> some View {
-        let single = audios.count == 1 ? audios.first : nil
-        animatableRow(label: "Volume", clipId: single?.id, property: .volume) {
-            ScrubbableNumberField(
-                value: sharedClipValue(audios) { clip in
-                    clip.liveVolumeKfDb(at: editor.activeFrame) ?? VolumeScale.dbFromLinear(clip.volume)
-                },
-                range: VolumeScale.floorDb...VolumeScale.ceilingDb,
-                format: "%.1f",
-                valueSuffix: " dB",
-                dragSensitivity: 0.3,
-                fieldWidth: 56,
-                displayTextOverride: { db in db <= VolumeScale.floorDb ? "-∞ dB" : nil },
-                onChanged: { db in
-                    for c in audios { editor.applyVolume(clipId: c.id, valueDb: db) }
-                }
-            ) { db in
-                commitToClips(audios, actionName: "Change Volume") { c in
-                    editor.commitVolume(clipId: c.id, valueDb: db)
-                }
-            }
-        }
-    }
-
-    @ViewBuilder
-    private func fadeRow(label: String, clips: [Clip], edge: FadeEdge) -> some View {
-        let fps = Double(max(1, editor.timeline.fps))
-        let single = clips.count == 1 ? clips.first : nil
-        let maxSeconds = single.map { Double($0.durationFrames) / fps } ?? 60.0
-        let actionName = edge == .left ? "Change Fade In" : "Change Fade Out"
-        propertyRow(label: label) {
-            ScrubbableNumberField(
-                value: sharedClipValue(clips) { clip in
-                    Double(clip.fadeFrames(edge)) / fps
-                },
-                range: 0...maxSeconds,
-                format: "%.2f",
-                valueSuffix: " s",
-                dragSensitivity: 0.02,
-                fieldWidth: 56,
-                onChanged: { seconds in
-                    let frames = Int((seconds * fps).rounded())
-                    for c in clips { editor.applyFade(clipId: c.id, edge: edge, frames: frames) }
-                }
-            ) { seconds in
-                let frames = Int((seconds * fps).rounded())
-                commitToClips(clips, actionName: actionName) { c in
-                    editor.commitFade(clipId: c.id, edge: edge, frames: frames)
-                }
-            }
-        }
-        .frame(height: KeyframesMetrics.rowHeight)
-    }
-
-
-    // MARK: - Effects Tab
-
-    private struct EffectControl: Hashable {
-        let effectId: String
-        let paramKey: String
-        var label: String? = nil
-        var gradient: [Color]? = nil
-    }
-
-    /// Basic › Tone.
-    private var toneControls: [EffectControl] {
-        [
-            EffectControl(effectId: "color.exposure", paramKey: "ev"),
-            EffectControl(effectId: "color.contrast", paramKey: "amount"),
-            EffectControl(effectId: "color.highlightsShadows", paramKey: "highlights"),
-            EffectControl(effectId: "color.highlightsShadows", paramKey: "shadows"),
-            EffectControl(effectId: "color.blacksWhites", paramKey: "blacks"),
-            EffectControl(effectId: "color.blacksWhites", paramKey: "whites"),
-        ]
-    }
-
-    /// Basic › White Balance.
-    private var whiteBalanceControls: [EffectControl] {
-        [
-            EffectControl(effectId: "color.temperature", paramKey: "temperature", gradient: AppTheme.Slider.tempGradient),
-            EffectControl(effectId: "color.temperature", paramKey: "tint", gradient: AppTheme.Slider.tintGradient),
-        ]
-    }
-
-    /// Basic › Presence.
-    private var presenceControls: [EffectControl] {
-        [
-            EffectControl(effectId: "color.vibrance", paramKey: "amount"),
-            EffectControl(effectId: "color.saturation", paramKey: "amount"),
-        ]
-    }
-
-    /// Blur/sharpen/stylize as single always-on rows, labeled by effect name.
-    private var stylizeControls: [EffectControl] {
-        [
-            EffectControl(effectId: "blur.gaussian", paramKey: "radius", label: "Blur"),
-            EffectControl(effectId: "blur.sharpen", paramKey: "amount", label: "Sharpen"),
-            EffectControl(effectId: "stylize.vignette", paramKey: "intensity", label: "Vignette"),
-        ]
-    }
-
-    /// Canonical order the fixed adjustment sections insert their effects in.
-    private var alwaysOnEffectOrder: [String] {
-        ["color.exposure", "color.contrast", "color.highlightsShadows", "color.blacksWhites",
-         "color.temperature", "color.vibrance", "color.saturation", "color.wheels", "color.curves",
-         "blur.gaussian", "blur.sharpen", "stylize.vignette"]
-    }
-
-    @ViewBuilder
-    private func effectsTabContent() -> some View {
-        let clips = nonTextVisualClips
-        VStack(alignment: .leading, spacing: AppTheme.Spacing.lg) {
-            capsuleTabBar(titles: AdjustTab.allCases.map(\.rawValue), selected: adjustSubTab.rawValue) { title in
-                if let tab = AdjustTab(rawValue: title) { adjustSubTab = tab }
-            }
-            switch adjustSubTab {
-            case .basic:
-                adjustmentSection(title: "Tone", controls: toneControls, clips: clips)
-                adjustmentSection(title: "White Balance", controls: whiteBalanceControls, clips: clips)
-                adjustmentSection(title: "Presence", controls: presenceControls, clips: clips)
-            case .color:
-                curvesSection(clips: clips)
-                wheelsSection(clips: clips)
-            case .effects:
-                adjustmentSection(title: "Effects", controls: stylizeControls, clips: clips)
-            }
-        }
-    }
-
-    // MARK: Curves
-
-    @ViewBuilder
-    private func curvesSection(clips: [Clip]) -> some View {
-        VStack(alignment: .leading, spacing: AppTheme.Spacing.md) {
-            HStack(spacing: AppTheme.Spacing.xs) {
-                sectionTitleLabel(title: "Curves")
-                Spacer()
-                if anyAdjusted(["color.curves"], clips: clips) {
-                    resetButton(
-                        onReset: { setCurve(GradeCurve(), clips: clips, commit: true, action: "Reset Curves") },
-                        help: "Reset curves"
-                    )
-                }
-            }
-            CurveEditorView(
-                curve: curve(in: clips.first?.effects ?? []),
-                onChange: { setCurveChannel($0, points: $1, clips: clips, commit: false, action: "Edit Curves") },
-                onCommit: { setCurveChannel($0, points: $1, clips: clips, commit: true, action: "Edit Curves") }
-            )
-            .padding(.leading, sectionContentIndent)
-        }
-    }
-
-    private func curve(in effects: [Effect]) -> GradeCurve {
-        guard let json = effects.first(where: { $0.type == "color.curves" })?
-            .params["curve"]?.string else { return GradeCurve() }
-        return GradeCurve(json: json) ?? GradeCurve()
-    }
-
-    private func setCurveChannel(
-        _ channel: CurveEditorView.Channel,
-        points: [CurvePoint],
-        clips: [Clip],
-        commit: Bool,
-        action: String
-    ) {
-        let mutate: (inout [Effect]) -> Void = { [self] effects in
-            var curve = curve(in: effects)
-            switch channel {
-            case .master: curve.master = points
-            case .red: curve.red = points
-            case .green: curve.green = points
-            case .blue: curve.blue = points
-            }
-            upsertCurve(curve, in: &effects)
-        }
-        if commit {
-            commitEffects(clips, actionName: action, mutate)
-        } else {
-            applyEffects(clips, mutate)
-        }
-    }
-
-    /// Upsert the curves effect in place (stable id), pruning it when the curve is identity.
-    private func setCurve(_ curve: GradeCurve, clips: [Clip], commit: Bool, action: String) {
-        let mutate: (inout [Effect]) -> Void = { [self] effects in
-            upsertCurve(curve, in: &effects)
-        }
-        if commit {
-            commitEffects(clips, actionName: action, mutate)
-        } else {
-            applyEffects(clips, mutate)
-        }
-    }
-
-    private func upsertCurve(_ curve: GradeCurve, in effects: inout [Effect]) {
-        let existing = effects.firstIndex { $0.type == "color.curves" }
-        guard !curve.isIdentity, let json = curve.encoded() else {
-            if let existing { effects.remove(at: existing) }
-            return
-        }
-        if let existing {
-            effects[existing].params["curve"] = EffectParam(string: json)
-        } else {
-            var effect = Effect(type: "color.curves")
-            effect.params["curve"] = EffectParam(string: json)
-            effects.insert(effect, at: alwaysOnInsertIndex(effects, for: "color.curves"))
-        }
-    }
-
-    // MARK: Color wheels
-
-    @ViewBuilder
-    private func wheelsSection(clips: [Clip]) -> some View {
-        VStack(alignment: .leading, spacing: AppTheme.Spacing.md) {
-            HStack(spacing: AppTheme.Spacing.xs) {
-                sectionTitleLabel(title: "Color Wheels")
-                Spacer()
-                if anyAdjusted(["color.wheels"], clips: clips) {
-                    HoldToPreviewButton(
-                        onPress: { previewSection(["color.wheels"], clips: clips, enabled: false) },
-                        onRelease: { previewSection(["color.wheels"], clips: clips, enabled: true) }
-                    )
-                    resetButton(
-                        onReset: { resetEffects(["color.wheels"], clips: clips, actionName: "Reset Color Wheels") },
-                        help: "Reset color wheels"
-                    )
-                }
-            }
-            HStack(alignment: .top, spacing: AppTheme.Spacing.md) {
-                wheelControl("Lift", prefix: "lift", clips: clips)
-                wheelControl("Gamma", prefix: "gamma", clips: clips)
-                wheelControl("Gain", prefix: "gain", clips: clips)
-            }
-            .frame(maxWidth: .infinity)
-            .padding(.leading, sectionContentIndent)
-        }
-    }
-
-    private func wheelControl(_ title: String, prefix: String, clips: [Clip]) -> some View {
-        let mKey = "\(prefix)_m"
-        let mSpec = EffectRegistry.descriptor(id: "color.wheels")?.params.first { $0.key == mKey }
-        let mDefault = mSpec?.defaultValue ?? 0
-        let mRange = mSpec?.range ?? 0...2
-        return ColorWheelControl(
-            title: title,
-            x: sharedClipValue(clips) { wheelParam($0, "\(prefix)_x", default: 0) } ?? 0,
-            y: sharedClipValue(clips) { wheelParam($0, "\(prefix)_y", default: 0) } ?? 0,
-            master: sharedClipValue(clips) { wheelParam($0, mKey, default: mDefault) } ?? mDefault,
-            masterRange: mRange,
-            masterDefault: mDefault,
-            onColorChanged: { setWheelColor(prefix, $0, $1, clips: clips, commit: false) },
-            onColorCommit: { setWheelColor(prefix, $0, $1, clips: clips, commit: true) },
-            onMasterChanged: { setControlParam(EffectControl(effectId: "color.wheels", paramKey: mKey), label: title, value: $0, clips: clips, commit: false) },
-            onMasterCommit: { setControlParam(EffectControl(effectId: "color.wheels", paramKey: mKey), label: title, value: $0, clips: clips, commit: true) }
-        )
-    }
-
-    private func wheelParam(_ clip: Clip, _ key: String, default def: Double) -> Double {
-        (clip.effects ?? []).first { $0.type == "color.wheels" }?.params[key]?.resolved(at: 0, default: def) ?? def
-    }
-
-    /// Both pad axes upserted in one mutation so a drag is a single undo entry.
-    private func setWheelColor(_ prefix: String, _ x: Double, _ y: Double, clips: [Clip], commit: Bool) {
-        let xc = EffectControl(effectId: "color.wheels", paramKey: "\(prefix)_x")
-        let yc = EffectControl(effectId: "color.wheels", paramKey: "\(prefix)_y")
-        let mutate: (inout [Effect]) -> Void = { [self] effects in
-            upsertControl(&effects, control: xc, value: x)
-            upsertControl(&effects, control: yc, value: y)
-        }
-        if commit {
-            commitEffects(clips, actionName: "Adjust \(prefix.capitalized)", mutate)
-        } else {
-            applyEffects(clips, mutate)
-        }
-    }
-
-    // MARK: Always-on adjustment sections (Color, Effects)
-
-    @ViewBuilder
-    private func adjustmentSection(title: String, controls: [EffectControl], clips: [Clip]) -> some View {
-        let ids = Set(controls.map(\.effectId))
-        VStack(alignment: .leading, spacing: AppTheme.Spacing.md) {
-            HStack(spacing: AppTheme.Spacing.xs) {
-                sectionTitleLabel(title: title)
-                Spacer()
-                if anyAdjusted(ids, clips: clips) {
-                    HoldToPreviewButton(
-                        onPress: { previewSection(ids, clips: clips, enabled: false) },
-                        onRelease: { previewSection(ids, clips: clips, enabled: true) }
-                    )
-                    resetButton(
-                        onReset: { resetEffects(ids, clips: clips, actionName: "Reset \(title)") },
-                        help: "Reset \(title.lowercased())"
-                    )
-                }
-            }
-            VStack(alignment: .leading, spacing: AppTheme.Spacing.md) {
-                ForEach(controls, id: \.self) { control in
-                    adjustmentRow(control, clips: clips)
-                }
-            }
-            .padding(.leading, sectionContentIndent)
-        }
-    }
-
-    @ViewBuilder
-    private func adjustmentRow(_ control: EffectControl, clips: [Clip]) -> some View {
-        if let descriptor = EffectRegistry.descriptor(id: control.effectId),
-           let spec = descriptor.params.first(where: { $0.key == control.paramKey }) {
-            let label = control.label ?? spec.label
-            HStack(spacing: AppTheme.Spacing.sm) {
-                Text(label)
-                    .font(.system(size: AppTheme.FontSize.sm))
-                    .foregroundStyle(AppTheme.Text.secondaryColor)
-                    .lineLimit(1)
-                    .frame(width: AppTheme.Slider.labelColumn, alignment: .leading)
-                AdjustSlider(
-                    value: sharedClipValue(clips) { controlValue($0, control, spec) } ?? spec.defaultValue,
-                    range: spec.range,
-                    gradient: control.gradient,
-                    defaultValue: spec.defaultValue,
-                    onChanged: { setControlParam(control, label: label, value: $0, clips: clips, commit: false) },
-                    onCommit: { setControlParam(control, label: label, value: $0, clips: clips, commit: true) }
-                )
-                ScrubbableNumberField(
-                    value: sharedClipValue(clips) { controlValue($0, control, spec) },
-                    range: spec.range,
-                    format: effectParamFormat(spec),
-                    valueSuffix: spec.unit.isEmpty ? "" : " \(spec.unit)",
-                    dragSensitivity: effectParamSensitivity(spec),
-                    fieldWidth: 50,
-                    onChanged: { setControlParam(control, label: label, value: $0, clips: clips, commit: false) }
-                ) { setControlParam(control, label: label, value: $0, clips: clips, commit: true) }
-            }
-            .frame(height: KeyframesMetrics.rowHeight)
-        }
-    }
-
-    private func controlValue(_ clip: Clip, _ control: EffectControl, _ spec: EffectParamSpec) -> Double {
-        (clip.effects ?? []).first { $0.type == control.effectId }?
-            .params[control.paramKey]?.resolved(at: 0, default: spec.defaultValue) ?? spec.defaultValue
-    }
-
-    private func setControlParam(
-        _ control: EffectControl, label: String, value: Double, clips: [Clip], commit: Bool
-    ) {
-        let mutate: (inout [Effect]) -> Void = { [self] effects in
-            upsertControl(&effects, control: control, value: value)
-        }
-        if commit {
-            commitEffects(clips, actionName: "Change \(label)", mutate)
-        } else {
-            applyEffects(clips, mutate)
-        }
-    }
-
-    /// Upsert one param into the singleton effect of its type, inserting in canonical
-    /// order when first touched and pruning it when every param returns to default
-    /// (so a neutral adjustment carries no effect / no render pass).
-    private func upsertControl(_ effects: inout [Effect], control: EffectControl, value: Double) {
-        guard let descriptor = EffectRegistry.descriptor(id: control.effectId) else { return }
-        if let i = effects.firstIndex(where: { $0.type == control.effectId }) {
-            effects[i].params[control.paramKey] = EffectParam(value: value)
-            let allDefault = descriptor.params.allSatisfy { spec in
-                (effects[i].params[spec.key]?.value ?? spec.defaultValue) == spec.defaultValue
-            }
-            if allDefault { effects.remove(at: i) }
-        } else {
-            let paramDefault = descriptor.params.first { $0.key == control.paramKey }?.defaultValue
-            guard value != paramDefault else { return }
-            var effect = descriptor.makeEffect()
-            effect.params[control.paramKey] = EffectParam(value: value)
-            effects.insert(effect, at: alwaysOnInsertIndex(effects, for: control.effectId))
-        }
-    }
-
-    private func alwaysOnInsertIndex(_ effects: [Effect], for effectId: String) -> Int {
-        let rank = alwaysOnEffectOrder.firstIndex(of: effectId) ?? Int.max
-        return effects.firstIndex { (alwaysOnEffectOrder.firstIndex(of: $0.type) ?? Int.max) > rank } ?? effects.count
-    }
-
-    private func anyAdjusted(_ ids: Set<String>, clips: [Clip]) -> Bool {
-        clips.contains { ($0.effects ?? []).contains { ids.contains($0.type) } }
-    }
-
-    private func resetEffects(_ ids: Set<String>, clips: [Clip], actionName: String) {
-        commitEffects(clips, actionName: actionName) { effects in
-            effects.removeAll { ids.contains($0.type) }
-        }
-    }
-
-    /// Live preview of a section toggled off/on. Both go through the cheap
-    /// refresh-visuals path (no full composition rebuild), so the release doesn't flicker.
-    private func previewSection(_ ids: Set<String>, clips: [Clip], enabled: Bool) {
-        applyEffects(clips) { effects in
-            for i in effects.indices where ids.contains(effects[i].type) {
-                effects[i].enabled = enabled
-            }
-        }
-    }
-
-    private func effectParamFormat(_ spec: EffectParamSpec) -> String {
-        (spec.range.upperBound - spec.range.lowerBound) <= 20 ? "%.2f" : "%.0f"
-    }
-
-    private func effectParamSensitivity(_ spec: EffectParamSpec) -> Double {
-        max(0.01, (spec.range.upperBound - spec.range.lowerBound) / 200)
-    }
-
-    /// Live edit (no undo entry) — mirrors applyClipProperty's refresh-only path.
-    private func applyEffects(_ clips: [Clip], _ mutate: @escaping (inout [Effect]) -> Void) {
-        editor.applyClipProperties(clipIds: clips.map(\.id)) { c in
-            var effects = c.effects ?? []
-            mutate(&effects)
-            c.effects = effects.isEmpty ? nil : effects
-        }
-    }
-
-    /// One undoable entry across all selected clips.
-    private func commitEffects(
-        _ clips: [Clip], actionName: String, _ mutate: @escaping (inout [Effect]) -> Void
-    ) {
-        editor.undoManager?.beginUndoGrouping()
-        editor.commitClipProperties(clipIds: clips.map(\.id)) { c in
-            var effects = c.effects ?? []
-            mutate(&effects)
-            c.effects = effects.isEmpty ? nil : effects
-        }
-        editor.undoManager?.endUndoGrouping()
-        editor.undoManager?.setActionName(actionName)
-    }
-
-    @ViewBuilder
-    private func speedSection(clips: [Clip]) -> some View {
+    func speedSection(clips: [Clip]) -> some View {
         if !clips.isEmpty {
             VStack(alignment: .leading, spacing: AppTheme.Spacing.smMd) {
                 sectionTitleLabel(title: "Playback")
@@ -862,7 +381,7 @@ struct InspectorView: View {
         }
     }
 
-    private func commitToClips(_ clips: [Clip], actionName: String, _ commit: (Clip) -> Void) {
+    func commitToClips(_ clips: [Clip], actionName: String, _ commit: (Clip) -> Void) {
         editor.undoManager?.beginUndoGrouping()
         for c in clips { commit(c) }
         editor.undoManager?.endUndoGrouping()
@@ -901,7 +420,7 @@ struct InspectorView: View {
 
     /// Property row with an optional keyframe stamp button after the value field.
     @ViewBuilder
-    private func animatableRow<Fields: View>(
+    func animatableRow<Fields: View>(
         label: String,
         clipId: String?,
         property: AnimatableProperty,
@@ -973,7 +492,7 @@ struct InspectorView: View {
     }
 
     /// Rows sit flush-left under their uppercase section header.
-    private var sectionContentIndent: CGFloat { 0 }
+    var sectionContentIndent: CGFloat { 0 }
 
     private func transformHeader(clips: [Clip]) -> some View {
         collapsibleHeader(
@@ -1087,7 +606,7 @@ struct InspectorView: View {
         }
     }
 
-    private func sectionTitleLabel(title: String) -> some View {
+    func sectionTitleLabel(title: String) -> some View {
         Text(title.uppercased())
             .font(.system(size: AppTheme.FontSize.xxs, weight: .semibold))
             .tracking(AppTheme.Tracking.wide)
@@ -1095,7 +614,7 @@ struct InspectorView: View {
             .fixedSize()
     }
 
-    private func resetButton(onReset: @escaping () -> Void, help: String?) -> some View {
+    func resetButton(onReset: @escaping () -> Void, help: String?) -> some View {
         Button(action: onReset) {
             Image(systemName: "arrow.counterclockwise")
                 .font(.system(size: AppTheme.FontSize.sm))
@@ -1107,7 +626,7 @@ struct InspectorView: View {
         .help(help ?? "Reset")
     }
 
-    private func propertyRow<Trailing: View>(
+    func propertyRow<Trailing: View>(
         label: String,
         @ViewBuilder trailing: () -> Trailing
     ) -> some View {
@@ -1408,7 +927,7 @@ struct InspectorView: View {
         return out
     }
 
-    private var selectedAudioClips: [Clip] {
+    var selectedAudioClips: [Clip] {
         guard !editor.selectedClipIds.isEmpty else { return [] }
         var out: [Clip] = []
         for track in editor.timeline.tracks {
