+++ Sources/PalmierPro/Inspector/Tabs/TextTab.swift
